tres deployment function now checks return values of intermediate operations

+++ pyot/models/tres.py
@@ -37,7 +37,7 @@ MEDIA_ROOT = settings.MEDIA_ROOT
 TFMT = settings.TFMT
 SERVER_ADDRESS = settings.SERVER_ADDRESS
 PROJECT_ROOT = settings.PROJECT_ROOT
-tmpDir = settings.TRES_PWN_SCRIPT_TMP 
+tmpDir = settings.TRES_PWN_SCRIPT_TMP
 
 WAIT_TIMEOUT = 30
 T_RES_TOOLS = settings.PROJECT_PATH + '/../t-res-tools/'
@@ -73,7 +73,7 @@ def verify_pf(sender, instance, raw, **kwargs):
     '''
     Validates the syntax of the processing function
     '''
-    print 'Validating '  + str(instance.sourcefile) + ' ...'
+    print 'Validating ' + str(instance.sourcefile) + ' ...'
     py_compile.compile(str(instance.sourcefile), doraise=True)
 
 pre_save.connect(verify_pf, sender=TResProcessing)
@@ -164,7 +164,7 @@ class TResT(models.Model):
     output = models.ForeignKey(Resource, related_name='OutputDestination',
                                null=True)
     TResResource = models.ForeignKey(Resource, null=True,
-                                     related_name='TresResource') 
+                                     related_name='TresResource')
     state = models.CharField(max_length=10, blank=False, choices=TRES_STATES,
                              default='CREATED')
     emu = models.ForeignKey(EmulatorState, null=True)
@@ -183,11 +183,11 @@ class TResT(models.Model):
 
         if t_res_resource.uri != '/tasks':
             raise Exception('TResResource must have /tasks uri')
-        self.TResResource = t_res_resource #FIXME only update if the installation is succesfull.
+        self.TResResource = t_res_resource # FIXME only update if the installation is succesfull.
 
 
-        #split the operation in phases, single tasks
-        #compile the script
+        # split the operation in phases, single tasks
+        # compile the script
         basename = os.path.basename(str(self.pf.sourcefile))
         compile_command = tresCompile + ' ' + tresPMfeat + ' ' + str(self.pf.sourcefile)
         p = subprocess.check_call([compile_command],
@@ -195,74 +195,78 @@ class TResT(models.Model):
                              stdout=subprocess.PIPE,
                              stderr=subprocess.PIPE,
                              shell=True, cwd=SCRIPT_FOLDER)
-        #start a task downloading the script from the server
+        # start a task downloading the script from the server
         pycFilename = basename + 'c'
-        tresDownloadScript.apply_async(args=[pycFilename],
+        r = tresDownloadScript.apply_async(args=[pycFilename],
                                        queue=t_res_resource.host.getQueue())
+        r.wait()
+        result = r.result
+        if result.code != SUCCESS:
+            return Response(FAILURE, 'PWN: Error downloading pyc.')
         try:
-            newTask = Resource.objects.get(host=self.TResResource.host, uri = '/tasks/'+self.pf.name)
+            newTask = Resource.objects.get(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
         except Resource.DoesNotExist:
-            newTask = Resource.objects.create(host=self.TResResource.host, uri = '/tasks/'+self.pf.name)
-    
+            newTask = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
+
         r = newTask.PUT()
-        print 'first put result = '+ r.code
+        print 'first put result = ' + r.code
         if r.code != CREATED:
             newTask.delete()
-            return 'Error creating new resource: ' + '/tasks/' + self.pf.name
+            return Response(FAILURE, 'Error creating new resource: ' + '/tasks/' + self.pf.name)
 
-        newIs = Resource.objects.create(host=self.TResResource.host, uri = '/tasks/' + self.pf.name+'/is')
-        newOd = Resource.objects.create(host=self.TResResource.host, uri = '/tasks/' + self.pf.name+'/od')
-        newPf = Resource.objects.create(host=self.TResResource.host, uri = '/tasks/' + self.pf.name+'/pf')
-        newLo = Resource.objects.create(host=self.TResResource.host, uri = '/tasks/' + self.pf.name+'/lo')    
-            
-        r =  newPf.PUT(inputfile=tmpDir + '/' + pycFilename, block=64)
-        print 'PF put result = '+ r.code
-    
+        newIs = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/is')
+        newOd = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/od')
+        newPf = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/pf')
+        _newLo = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/lo')
+
+        r = newPf.PUT(inputfile=tmpDir + '/' + pycFilename, block=64)
+        print 'PF put result = ' + r.code
+
         if r.code != CHANGED:
-            return 'Error uploading processing function.'     
+            return Response(FAILURE, 'Error uploading processing function.')
 
         if self.output:
-            r = newOd.PUT(payload='<'+ self.output.getFullURI() +'>')
-            print 'OD put result = '+ r.code
+            r = newOd.PUT(payload='<' + self.output.getFullURI() + '>')
+            print 'OD put result = ' + r.code
             if r.code != CHANGED:
-                return 'Error updating OD resource'
+                return Response(FAILURE, 'Error updating OD resource')
         for inp in self.inputS.all():
-            r = newIs.POST(payload='<'+ inp.getFullURI() +'>')
-            print 'IS put result = '+ r.code
+            r = newIs.POST(payload='<' + inp.getFullURI() + '>')
+            print 'IS put result = ' + r.code + ' ' + inp.getFullURI()
             if r.code != CHANGED:
-                return 'Error updating OD resource'
-        
+                return Response(FAILURE, 'Error updating IS resource')
+
         self.state = 'INSTALLED'
-        self.save() #update TResResource and state
-        
-        return 'Tres Task ' + str(self) + ' Installed'
+        self.save() # update TResResource and state
+
+        return Response(SUCCESS, 'Tres Task ' + str(self) + ' Installed')
 
     def uninstall(self):
         """
         Uninstalls a T-Res task from a node.
         """
-        newTask = Resource.objects.get(host=self.TResResource.host, uri = '/tasks/' + self.pf.name)
+        newTask = Resource.objects.get(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
         r = newTask.DELETE()
         if r.code == DELETED:
             newTask.delete()
-            Resource.objects.filter(host=self.TResResource.host, 
-                                    uri__startswith='/tasks/'+self.pf.name).delete()
+            Resource.objects.filter(host=self.TResResource.host,
+                                    uri__startswith='/tasks/' + self.pf.name).delete()
             self.state = 'CLEARED'
-            self.save()                                    
-            return 'Task ' + self.pf.name+ ' uninstalled' 
+            self.save()
+            return Response(SUCCESS, 'Task ' + self.pf.name + ' uninstalled')
         else:
-            return 'Error uninstalling task ' + self.pf.name       
-        
+            return Response(FAILURE, 'Error uninstalling task ' + self.pf.name)
+
     def start(self):
         '''
         Activates a TRes Task by sending a POST request to the task resource.
         '''
         if self.state == 'RUNNING':
-            return 'Task is already running'
+            return Response(SUCCESS, 'Task is already running')
         Pf = Resource.objects.get(host=self.TResResource.host,
-                                  uri='/tasks/'+self.pf.name)
-        r = Pf.POST()
-        print r.content, r.code
+                                  uri='/tasks/' + self.pf.name)
+        r = Pf.POST(query="op=on")
+        print str(self.TResResource), r.content, r.code
         if r.content == "Task now running":
             self.state = 'RUNNING'
             self.save()
@@ -273,11 +277,11 @@ class TResT(models.Model):
         Stops the execution of the T-Res Task.
         """
         if self.state != 'RUNNING':
-            return 'Task is not running'
+            return Response(SUCCESS, 'Task is not running')
         Pf = Resource.objects.get(host=self.TResResource.host,
-                                  uri='/tasks/'+self.pf.name)
-        r = Pf.POST()
-        print r.content, r.code
+                                  uri='/tasks/' + self.pf.name)
+        r = Pf.POST(query="op=off")
+        print str(self.TResResource), r.content, r.code
         if r.content == "Task now halted":
             self.state = 'STOPPED'
             self.save()
@@ -296,7 +300,7 @@ class TResT(models.Model):
         retrieved or observed.
         """
         last_output = Resource.objects.get(host=self.TResResource.host,
-                                           uri='/tasks/'+self.pf.name+'/lo')
+                                           uri='/tasks/' + self.pf.name + '/lo')
         return last_output
 
     def getInputSource(self):
@@ -304,7 +308,7 @@ class TResT(models.Model):
         Returns the list of Input Source resources.
         """
         _is = Resource.objects.get(host=self.TResResource.host,
-                                   uri='/tasks/'+self.pf.name+'/is')
+                                   uri='/tasks/' + self.pf.name + '/is')
         return _is
 
     def getOutputDestination(self):
@@ -312,7 +316,7 @@ class TResT(models.Model):
         Returns the list of Output Destination resources.
         """
         _od = Resource.objects.get(host=self.TResResource.host,
-                                   uri='/tasks/'+self.pf.name+'/od')
+                                   uri='/tasks/' + self.pf.name + '/od')
         return _od
 
     def emulate(self, duration=120):
@@ -354,7 +358,7 @@ class TResT(models.Model):
         self.emu.save()
         self.save()
         #from cStringIO import StringIO
-        sys.path.append(PROJECT_ROOT+'/tres')
+        sys.path.append(PROJECT_ROOT + '/tres')
         sys.argv = [self.id]
         #oldio = (sys.stdin, sys.stdout, sys.stderr)
         #sio = StringIO()
